requirements.txt
fastapi
sqlmodel
uvicorn
passlib
python-dotenv
psycopg2-binary
pydantic[email]
bcrypt
python-jose
fastapi-jwt-auth